Resolves the decisions.md conflict by renumbering the maplibre v6 entry
331 -> 347 behind the ADRs that landed on main. Lockfiles and
apps/web/package.json auto-merged; verified maplibre-gl resolves to
6.0.0 and fit-file-parser 4.0.1 (#630) survived in both lockfiles.

+++ apps/custom_watch/core/src/settings_queue.rs
@@ -103,7 +103,7 @@ mod tests {
         }
     }
 
-    fn page_mask(mask: u32) -> WatchSettings {
+    fn page_mask(mask: u64) -> WatchSettings {
         WatchSettings {
             pages: Some(mask),
             ..WatchSettings::default()
